+++ backend/windmill-worker/src/bash_executor.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap, process::Stdio};
+use std::{collections::HashMap, fs, process::Stdio};
 
 use itertools::Itertools;
 use regex::Regex;
@@ -11,13 +11,17 @@ use windmill_queue::CanceledBy;
 const BIN_BASH: &str = "/bin/bash";
 const NSJAIL_CONFIG_RUN_BASH_CONTENT: &str = include_str!("../nsjail/run.bash.config.proto");
 
+lazy_static::lazy_static! {
+    static ref RE_POWERSHELL_IMPORTS: Regex = Regex::new(r#"^Import-Module\s+(?:-Name\s+)?"?([^-\s"]+)"?"#).unwrap();
+}
+
 use crate::{
     common::{
         build_args_map, get_reserved_variables, handle_child, read_file, read_file_content,
         set_logs, start_child_process, write_file,
     },
     AuthedClientBackgroundTask, DISABLE_NSJAIL, DISABLE_NUSER, HOME_ENV, NSJAIL_PATH, PATH_ENV,
-    TZ_ENV,
+    POWERSHELL_CACHE_DIR, TZ_ENV,
 };
 
 lazy_static::lazy_static! {
@@ -174,8 +178,6 @@ pub async fn handle_powershell_job(
     worker_name: &str,
     envs: HashMap<String, String>,
 ) -> Result<Box<RawValue>, Error> {
-    logs.push_str("\n\n--- POWERSHELL CODE EXECUTION ---\n");
-    set_logs(logs, &job.id, db).await;
     let pwsh_args = {
         let args = build_args_map(job, client, db).await?.map(Json);
         let job_args = if args.is_some() {
@@ -205,6 +207,90 @@ pub async fn handle_powershell_job(
     let content = content
         .replace('$', r"\$") // escape powershell variables
         .replace("`", r"\`"); // escape powershell backticks
+
+    let installed_modules = fs::read_dir(POWERSHELL_CACHE_DIR)?
+        .filter_map(|x| {
+            x.ok().map(|x| {
+                x.path()
+                    .display()
+                    .to_string()
+                    .split('/')
+                    .last()
+                    .unwrap_or_default()
+                    .to_lowercase()
+            })
+        })
+        .collect::<Vec<String>>();
+
+    let mut install_string: String = String::new();
+    for line in content.lines() {
+        for cap in RE_POWERSHELL_IMPORTS.captures_iter(line) {
+            let module = cap.get(1).unwrap().as_str();
+            if !installed_modules.contains(&module.to_lowercase()) {
+                logs.push_str(&format!("\n{} not found in cache", module.to_string()));
+                // instead of using Install-Module, we use Save-Module so that we can specify the installation path
+                install_string.push_str(&format!(
+                    "Save-Module -Path {} -Force {};",
+                    POWERSHELL_CACHE_DIR, module
+                ));
+            } else {
+                logs.push_str(&format!("\n{} found in cache", module.to_string()));
+            }
+        }
+    }
+    set_logs(logs, &job.id, db).await;
+
+    if !install_string.is_empty() {
+        logs.push_str("\n\nInstalling modules...");
+        set_logs(logs, &job.id, db).await;
+        let child = Command::new("pwsh")
+            .args(&["-Command", &install_string])
+            .stdout(Stdio::piped())
+            .stderr(Stdio::piped())
+            .spawn()?;
+
+        handle_child(
+            &job.id,
+            db,
+            logs,
+            mem_peak,
+            canceled_by,
+            child,
+            false,
+            worker_name,
+            &job.workspace_id,
+            "powershell install",
+            job.timeout,
+            false,
+        )
+        .await?;
+    }
+
+    logs.push_str("\n\n--- POWERSHELL CODE EXECUTION ---\n");
+    set_logs(logs, &job.id, db).await;
+
+    // make sure default (only allhostsallusers) modules are loaded, disable autoload (cache can be large to explore especially on cloud) and add /tmp/windmill/cache to PSModulePath
+    let profile = format!(
+        "\\$PSModuleAutoloadingPreference = 'None'
+\\$PSModulePathBackup = \\$env:PSModulePath
+\\$env:PSModulePath = (\\$Env:PSModulePath -split ':')[-1]
+Get-Module -ListAvailable | Import-Module
+\\$env:PSModulePath = \"{}:\\$env:PSModulePathBackup\"",
+        POWERSHELL_CACHE_DIR
+    );
+    // make sure param() is first
+    let param_match = windmill_parser_bash::RE_POWERSHELL_PARAM.find(&content);
+    let content: String = if let Some(param_match) = param_match {
+        let param_match = param_match.as_str();
+        format!(
+            "{}\n{}\n{}",
+            param_match,
+            profile,
+            content.replace(param_match, "")
+        )
+    } else {
+        format!("{}\n{}", profile, content)
+    };
 
     write_file(job_dir, "main.sh", &format!("set -e\ncat > script.ps1 << EOF\n{content}\nEOF\npwsh -File script.ps1 {pwsh_args}\necho \"\"\nsleep 0.02")).await?;
     let token = client.get_token().await;